git rm proxy files
